refactor: 내 설문지 페이지 코드 분리
